+++ Simple-Projects/Simple_Projects/Pen&Paper/Pen.h
@@ -1,20 +1,21 @@
 #ifndef PEN_H
 #define PEN_H
 
+#include <iostream>
 #include "Paper.h"
 
 class OutOfInk {};
 
 class Pen {
     private:
-        size_t inkAmount;
-        size_t inkCapacity;
+        int inkAmount;
+        int inkCapacity;
     public:
-        Pen(size_t inkCapacity=4096);
+        Pen(int inkCapacity=4096);
         ~Pen();
 
-        size_t getInkAmount() const;
-        size_t getInkCapacity() const;
+        int getInkAmount() const;
+        int getInkCapacity() const;
 
         void write(Paper& paper, const std::string& message);
         void refill();
